API-SERVICE/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
asyncssh==2.12.0
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
ecdsa==0.18.0
elasticsearch==7.17.6
fastapi==0.83.0
frozenlist==1.3.1
h11==0.14.0
idna==3.4
multidict==6.0.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
PyJWT==2.6.0
python-jose==3.3.0
pytz==2022.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.16.0
yarl==1.8.1
